Add changelog and some cleanups
