+++ server/src/test/java/io/crate/integrationtests/BaseRolesIntegrationTest.java
@@ -27,12 +27,6 @@ import org.elasticsearch.test.IntegTestCase;
 import org.junit.After;
 import org.junit.Before;
 
-import io.crate.session.Session;
-import io.crate.session.Sessions;
-import io.crate.auth.Protocol;
-import io.crate.protocols.postgres.ConnectionProperties;
-import io.crate.role.Role;
-import io.crate.role.Roles;
 import io.crate.testing.SQLResponse;
 
 public abstract class BaseRolesIntegrationTest extends IntegTestCase {
@@ -66,16 +60,6 @@ public abstract class BaseRolesIntegrationTest extends IntegTestCase {
 
     public SQLResponse executeAsSuperuser(String stmt, Object[] args) {
         return execute(stmt, args);
-    }
-
-    public SQLResponse executeAs(String stmt, String userName) {
-        Sessions sqlOperations = cluster().getInstance(Sessions.class);
-        Roles roles = cluster().getInstance(Roles.class);
-        Role user = roles.getUser(userName);
-        try (Session session = sqlOperations.newSession(
-            new ConnectionProperties(null, null, Protocol.HTTP, null), null, user)) {
-            return execute(stmt, null, session);
-        }
     }
 
     protected void assertUserIsCreated(String userName) {
